Fix bug in HashMap.IterateChannel()
The function was exiting and closing the channel before the go routine finished writing data to it - only detectable once the unit test was in place to write enough values into the channel to take too long to get it done before the channel close. One point for unit tests.