Improve and fix logic issues with Config dereferencing
Also creates a bunch more Crazy log output during the configuration stage which was instrumental to locating the logic issue. We have added a DerefernceAll method to Config which allows us to apply multiple layers/passes to Config dereferencing in a single call (code reduction). Actually applying this highlighted a logic error in the original Config.Dereference() method which was dereferencing a configuration against ITSELF instead of against the one that was passed to it - OOPS! This is the kind of thing that should have been caught by a unit test. Too bad we don't have test coverage here - grr, tech debt already!