Add feature parity to NamedObjectstore suport in ObjectStoreManager
Now you can do all the same operations using either Scan or Name method of ObjectStore selection within the Manager's collection.

Next up: add more interesting capabilities such as retrieving ObjectCollection's with ObjectStoreManager (which is a precursor to adding support for MySQL ObjectStore where it will need to be able to return ObjectCollections as a result, not just a single Object.