WIP successful pull of MySQL data, but needs capture
We are reading data from mysql row results with Scan() now, but each row is overwriting the previous so that every row result record is the same as the last one. This is prsumably due to the use of pointers. We need a factory to pop out fresh records that are ready for writing...