WIP Back to factory pattern, but reflection still not right for cloning results
Getting close, but this accepted answer seems most important: https://stackoverflow.com/questions/40512323/golang-cast-interface-back-to-its-original-type

The problems is that we want to clone, as a factory, DTO structs with any member, which means they are an interface{}. But from reflection's view, an interface does not have any actual members, only methods that can be called. So when we use reflect to get fields, it's using a little magic to examine those fields but we cannot access them directly. The above example may have some working code to get fields from one interface ad set them in another, but unclear, needs experimentation