WIP - Still struggling with reflection in ResultFactory class
We may need to switch to the simplest form possible of a per result class method to get a replica of the class, and its property pointers. We might be able to get away with some boilerplate code in a nested struct/interface but the wrapper remains in the original class so that it knows what its own type is and can assert itself as needed, etc.