Add support for ANY request method
Normally we fan out module controller requests according to which request method is used and which method(s) are supported by the endpoint. The ANY method allows us to fan back in to a single handler for endpoints with the ANY method. We will still favor the other, more specialized endpoints where there is an exact match, but where there is no exact handler match, ANY applies! It's a catchall which, if present/implemented in the endpoint - WILL catch them all. This means the endpoint may also ONLY support this method to support ALL methods in the simgle handler where appropriate.