Refactor Logger to accept variadic input same as fmt.Sprintf()
This eliminates the need for many calls to include boilerplate fmt.Sprintf() wrapper to get a message logged - Logger now does this for them. It is backward compatible, so refactoring callers to eliminate the fmt.Sprintf() wrappers that they have in place is an optional refactoring (i.e. it won't break either way)