WIP - Refactoring ResultRow into an extensible struct
In the name of "return structs", we convert the map type of ResultRow into a struct which would enable us to extend the functionality of ResultRow's interface in the future, including by way of adding properties to the struct, in a way that is transparent to consumers (i.e. minimize the blast radius of such future changes).