WIP - Refactoring Cache
There are some strange overlapping design concerns between the ordered age list, the actual cache which holds the binary value, and elements of the age list that map to a cache entry key. Need to streamline the age list so that it is just a ordered list of keys that plucks a given key out of the set and moves it back to the front every time it is "touched" (to keep it from aging out). This list should not contain extra copies/references to the actual cache entry content.