WIP -Refactoring - DB - Test Coverage - BROKEN
Something is screwy with the Expect**() results around transactions and prepared statements. The mock class seems to have basic support for primitives, but not complete support for normal transactional/prepared statements. We might have to analyze their code and fix something... or else use theirs as a model to spin our own which is more complete. It's not even that complicated, so it's a little peculiar of a position to be in here...