WIP - Refactoring - DB - Move prepared statements behind the scenes
Previous commit left a broken test. It was because we didn't understand that calling Tx.Stmt(stmt) causes an already prepared statement to be prepared again in the context of the transaction... and this was not mocked. This called into question the value of calling Tx.Stmt(stmt) to prepare a statement for the transaction... which was already prepared for the transaction! Instead of leaving it to the caller to manage all this, we track transaction-specific and non-transacton prepared statements in a new structure now which allows the caller to simply supply the query and arguments and benefit from the protections provided by prepared statements without concern to transacitonality, nor to managing prepared statements at the application/caller layer.