WIP - Refactoring - Config - Rename methods conflicting with HashMap
When you embed interfaces, the embedded interface cannot have the same method names as the embedding struct. With Config acting as an extension of HashMap, but still wanting to override certain method behaviors that are more meaningful as a Config than as a HashMap, because Go has no concept of override, we simply rename the Config methods so that they don't collide. There is a way to make both work by producing an interface that represents the intersection between the two, potentially, but they would have to agree on the method signature which, in this case we don't. So we just dodge the problem entirely here.