WIP - Refactoring - Config - Add a GetBool() method
This new methods tries a few different ways to consider the "truthiness" of the value stored for a configuration key. If there is an obvious false (like empty string or nil/missing key) then we return that. Otherwise, try matching some common stringy variants of true booleans followed by non-0 integer boolean, and finally give up with a false.