WIP - Refactoring - DB - MySQL - Make everything use SQLQueryIfc
Don't resolve the SQL query into a flat string until the last possible moment, just ahead of sending it to the MySQL driver. This eliminates the need for rosolving tokens, etc in many places so that it only neds to be done in one place. It also gives us programmatic access/control to whatever dynamic capabilities are part of the SQLQuery implementation right up until it is ready to execute.