WIP - DB > MySQL > Nullables tset coverage
TODO: Revisit the use of the "Valid" flag in Scanner Interface for each nullable type as well as wrapper class and nullables base which maybe needs to check this instead of (or in addition to?) nullableType. It is all well and good to get/set programmatically through the normal interface here, but Scanner is another interface that will pull data from DB Results and we need to make sure that these are aligned. Valid is specifically what scanner uses to indicate null on those sql.null* types, so we had better be paying attention to it...