Add HashMap.ToLog() to dump current contents to Logger
We used to have a hacky version of this called Dump() which was good for early testing, but not very flexible. ToLog() now leverages some of the interface conventions that we have building out and provides flexibility to the caller to describe the structure being logged with a label as well as the LogLevel.