WIP - Add a MemcacheClient library
Using a Google open source client library project as the based. Code appears operable and is fiarly striaghtforward, but a bit disorganized, difficult to read. There is some good thinking around mutex locks for multithreaded concurrency. We may end up rewriting this ourselves so that we are not bound to their license, and the code is actually legible, maintainable, well organized, commented, configurable, etc, etc.