WIP - Add a ByteMap ([]byte variant of HashMap which is string oriented)
This should help us more reliably retain BLOB's and their ilk and still convert to string as needed. We may also want to just go for an interface{} map, but then the consumer would need to know what to cast each interface{} value to, so that doesn't help reduce boilerplate so much. Is there are way to achieve this with generics? Maybe? No?