WIP - Get example code working
Still a few TODO's remain; these are important to think through because the entire point of this exercise is to standardize our approach to dependency injection and reduce the amount of boilerplate code that ends up in our applicatoin layers. The more that can be baked into the library, the better. The biggest challenge WRT elimination of boilerplate seems to be the handling of of interface assertions which don't support dynamic typing (i.e. you can't pass the name of an interface as a string and get what you want out of it)