Complete build/test fixes for DB/MySQL
Note that ResultRow was using Nullable as a value rather than as a pointer to the value as returned by NewNullable(). This is important to remain consistent. The encoding/json.Marshaler Interface can only be implemented on either a pointer receiver or an object receiver, not both. Thus if NewNullable() implements as a pointer receiver, but then consumer dereferences the pointer, MarshalJSON can no longer see the pointer receiver interface and therefore does not invoke which results in munged JSON output.