Return empty set instead of nilable pointer for []dependencyInstance
nilable pointer implied that it maybe could be nil sometime. but it can't - it's never nil. it's always at least an empty set, and at best a populated set. possibly this was a weak attempt at not passing data structures bacby sending only the pointer, however this causes confusion in the consumer.