WIP - Notes on Json lexer, UTF-8 encoding, runes, etc.
Note: can we just draw a line and support UTF-8 only, all others must convert to use this? This represents an opinionated aspect of the platform that can reduce complexity by supporting more than necessary for all the different types of multi-byte character encoding: caller's responsibility to detect and convert as needed.