Refactor DataValue ToString() and ToJson() to common function
The only difference in the two encodings is whether strings are quoted. both need to represent the same data structures, so it only makes sense to account for both in the same way. But for ToString() output, it's important to be able to pull the bare string value unquoted so that the consumer needn't take the extra step and waste cycle to un-quote/unescape it. But it is generally anticipated that ToJson() will likely the most commonly used form