WIP - Fill in test coverage for ConfigItem
Switching up the function prototype for ValidateFunc to return error instead of bool vs our original GoLib/Config definition. The original was a tripping point on the implementation of a Configurable structure nearly ever time because Capture() returned errots, but Validate() returned bools - which was which required a lookup every time. And why wouldn't we want a validator to return the reason for validation failure? There could be any number of reasons for validation to fail, and now it has a voice to communicate that reason.