WIP - Added Clone factory for DataValue
That needs some test coverage as well as Config which is now using it.

I was looking to see if we could make this thing immutable entirely, which we could if we made the Config.DataValue a non-exported private member property and then replicated the DataValue interface on Config to pass calls on to the private member - but this is annoying to maintain and would not automatically pick up new methods of DataValue without propagating the pass-through function for every change/addition (plus it would break the build because Config would not be Interface-compatible with DataValue every time a change is made. So let's just resolve to put some cautionary notes in place to encourage consumers to make copies of Config that they pass to untrusted code/consumers and don't count on them being pristine afterward. Then if a consumer tampers, it's their own problem, it doesn't propagate back up to the caller/provider