WIP - Cache has at least two test cases related to pruning intermittently failing
Look for some sort of time-based race condition; and check out our time sync fix in the previous commit to see if the same fix is viable. It begs the question of whether a time based conditoin would manifest under real world conditions and end up with surprisng results? Perhaps the timing race condition needs to be fixed in the implementation and not in the test case, if this is truly the case.