WIP - MAYBE fixed all the test/logic issues here
We were doing strange things with pruning operations when calling from set() as well as from tests. Pruning functions accepted arguments suggesting that the future size would be something based on an addition, however the addition was already made before pruning was invoked, so there was math confusion dealing with future projection against something that already happened. This may have been the result of an earlier implementation that was aiming to prevent the data structure from ever exceeding the limit vs allowing it to exceed briefly, but then immediately pruning back to limits afterward.